+++ database/migrations/2021_02_12_230505_create_balances_table.php
@@ -15,7 +15,7 @@ class CreateBalancesTable extends Migration
     {
         Schema::create('balances', function (Blueprint $table) {
             $table->id();
-            $table->float('balance');
+            $table->bigInteger('balance');
             $table->timestamps();
         });
     }
